>IP JWT integration V0.4  solve issue with loadingDB, was CORS (added Auth header)
Need to change how users are loaded to only load the user itself from the user table.

Need to add the headers for saving the changes.

For log out, need to reset AuthToken.
